+++ src/components/pages/Checkout/Stepper.js
@@ -10,35 +10,35 @@ const stepperData = [
   {
     id: 1,
     icon: (
-      <FaShoppingBasket className="text-xl font-bold text-center md:text-2xl lg:text-3xl text-secbg-secondgraphy group-hover:text-white" />
+      <FaShoppingBasket className="text-lg font-bold text-center md:text-xl lg:text-2xl text-secbg-secondgraphy group-hover:text-white" />
     ),
     title: "Basket",
   },
   {
     id: 2,
     icon: (
-      <IoCloudUploadOutline className="text-xl font-bold text-center md:text-2xl lg:text-3xl text-secbg-secondgraphy group-hover:text-white" />
+      <IoCloudUploadOutline className="text-lg font-bold text-center md:text-xl lg:text-2xl text-secbg-secondgraphy group-hover:text-white" />
     ),
     title: "Upload Artwork",
   },
   {
     id: 3,
     icon: (
-      <TbTruckDelivery className="text-xl font-bold text-center md:text-2xl lg:text-3xl text-secbg-secondgraphy group-hover:text-white" />
+      <TbTruckDelivery className="text-lg font-bold text-center md:text-xl lg:text-2xl text-secbg-secondgraphy group-hover:text-white" />
     ),
     title: "Delivery ",
   },
   {
     id: 4,
     icon: (
-      <BsCashCoin className="text-xl font-bold text-center md:text-2xl lg:text-3xl text-secbg-secondgraphy group-hover:text-white" />
+      <BsCashCoin className="text-lg font-bold text-center md:text-xl lg:text-2xl text-secbg-secondgraphy group-hover:text-white" />
     ),
     title: " Payment",
   },
   {
     id: 5,
     icon: (
-      <MdOutlineAutoDelete className="text-xl font-bold text-center md:text-2xl lg:text-3xl text-secbg-secondgraphy group-hover:text-white" />
+      <MdOutlineAutoDelete className="text-lg font-bold text-center md:text-xl lg:text-2xl text-secbg-secondgraphy group-hover:text-white" />
     ),
     title: "Order seccess",
   },
