+++ src/components/pages/invoice/DownloadInvoice.js
@@ -1,31 +1,12 @@
 import React from "react";
 import jsPDFInvoiceTemplate from "jspdf-invoice-template";
-import {} from "jspdf-invoice-template";
-import { IoPrint } from "react-icons/io5";
 
-const PrintInvoice = () => {
-  function printPDF(pdfUrl) {
-    if (window !== undefined) {
-      const iframe = document.createElement("iframe");
-      iframe.style.display = "none";
-      iframe.src = pdfUrl;
-      document.body.appendChild(iframe);
-      iframe.contentWindow.print();
+import { FaDownload } from "react-icons/fa";
 
-      /* const newWindow = window.open(pdfUrl, "_blank");
-      if (newWindow) {
-        newWindow.onload = function () {
-          newWindow.print();
-        };
-      } else {
-        alert("Please allow popups for this website");
-      } */
-    }
-  }
+const DownloadInvoice = () => {
   const downloadInvoiceFile = () => {
     const invoiceData = {
-      outputType: "blob", // save, string, datauristring
-
+      outputType: "save", // save, string, datauristring
       returnJsPDFDocObject: true,
       fileName: "Invoice 2021",
       orientationLandscape: false,
@@ -128,24 +109,16 @@ Qty: 5000, Paper Type: 130gsm Gloss Finish Paper, Size: A4 Landscape,Folding: Ro
       pageLabel: "Page ",
     };
 
-    if (window !== undefined) {
-      const url = jsPDFInvoiceTemplate(invoiceData);
-
-      const print = URL.createObjectURL(url.blob);
-      printPDF(print);
-      /* jsPDFInvoiceTemplate(invoiceData, (pdf) => {
-        const pdfBlob = pdf.output("blob");
-        const pdfUrl = URL.createObjectURL(pdfBlob);
-        printPDF(pdfUrl);
-      }); */
+    if (typeof window !== "undefined") {
+      jsPDFInvoiceTemplate(invoiceData);
     }
   };
 
   return (
-    <button onClick={downloadInvoiceFile}>
-      <IoPrint />
+    <button>
+      <FaDownload />
     </button>
   );
 };
 
-export default PrintInvoice;
+export default DownloadInvoice;
